+++ onnxscript/optimizer/__init__.py
@@ -12,8 +12,8 @@ from onnxscript import ir, rewriter
 from onnxscript.optimizer import _constant_folding, _inliner
 from onnxscript.optimizer._legacy.constant_folding import fold_constants
 from onnxscript.optimizer._remove_unused import remove_unused_nodes
-from onnxscript.optimizer.remove_unused_function import remove_unused_functions
-from onnxscript.optimizer.simple_function_folding import (
+from onnxscript.optimizer._remove_unused_function import remove_unused_functions
+from onnxscript.optimizer._legacy._simple_function_folding import (
     inline_functions_with_unused_outputs,
     inline_simple_functions,
 )
