+++ onnxscript/optimizer/_function_folding_test.py
