+++ onnxscript/optimizer/_legacy/_simple_function_folding.py
