+++ onnxscript/optimizer/_legacy/_simple_function_folding_test.py
@@ -6,7 +6,8 @@ import unittest
 
 import onnx
 
-from onnxscript.optimizer import remove_unused_function, simple_function_folding
+from onnxscript.optimizer import _remove_unused_function
+from onnxscript.optimizer._legacy import _simple_function_folding
 
 
 class SingleNodeFunctionFoldingTest(unittest.TestCase):
@@ -32,8 +33,8 @@ foldable (x) => (return_val)
             """
         )
 
-        simple_function_folding.inline_simple_functions(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_simple_functions(model)
+        model = _remove_unused_function.remove_unused_functions(model)
 
         self.assertEqual(len(model.functions), 0)
 
@@ -59,8 +60,8 @@ foldable <dim>(x, y) => (return_val)
             """
         )
 
-        simple_function_folding.inline_simple_functions(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_simple_functions(model)
+        model = _remove_unused_function.remove_unused_functions(model)
 
         self.assertEqual(len(model.functions), 0)
         self.assertFalse(model.graph.node[0].attribute[0].ref_attr_name)
@@ -98,8 +99,8 @@ non_foldable <axis>(x, y) => (return_val)
             """
         )
 
-        simple_function_folding.inline_simple_functions(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_simple_functions(model)
+        model = _remove_unused_function.remove_unused_functions(model)
 
         self.assertEqual(len(model.functions), 1)
         self.assertEqual(model.functions[0].node[0].op_type, "Concat")
@@ -127,8 +128,8 @@ prim_cast <to>(x) => (return_val)
 }
             """
         )
-        simple_function_folding.inline_simple_functions(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_simple_functions(model)
+        model = _remove_unused_function.remove_unused_functions(model)
         self.assertEqual(len(model.functions), 0)
         self.assertEqual(len(model.graph.node), 3)
         self.assertEqual(model.graph.node[0].attribute[0].i, 10)
@@ -170,8 +171,8 @@ foldable_func (x, y) => (z_6)
              """
         )
 
-        simple_function_folding.inline_simple_functions(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_simple_functions(model)
+        model = _remove_unused_function.remove_unused_functions(model)
 
         self.assertEqual(len(model.functions), 0)
         self.assertEqual(len(model.graph.node), 2)
@@ -211,8 +212,8 @@ non_foldable <axis>(x, y) => (return_val)
             """
         )
 
-        simple_function_folding.inline_functions_with_unused_outputs(model)
-        model = remove_unused_function.remove_unused_functions(model)
+        _simple_function_folding.inline_functions_with_unused_outputs(model)
+        model = _remove_unused_function.remove_unused_functions(model)
         self.assertEqual(len(model.functions), 1)
