+++ onnxscript/optimizer/_remove_unused_function.py
